downloadable and download functionality improved.
downloadable can now take a job id as an argument, and will list all
the objects that are associated with that RDA batch job.

download can now take the path returned form downloadble and just
download that object.  In reality it will download all objects that
match the given prefix, so power users can profit from that.